+++ src/foundry/client/pixi/index.d.ts
@@ -1,12 +1,10 @@
 import "./board";
 import "./core";
 import "./groups";
-import "./layer";
 import "./layers";
 import "./perception";
 import "./placeable";
 import "./placeables";
-import "./placeables/index";
 import "./sources";
 import "./webgl";
 import "./workers";
