+++ src/foundry/client/pixi/layers/base/index.d.ts
@@ -1,0 +1,3 @@
+import "./canvas-layer";
+import "./interaction-layer";
+import "./placeables-layer";
