+++ src/foundry/client/pixi/layers/base/interaction-layer.d.ts
@@ -1,0 +1,8 @@
+export {};
+
+declare global {
+  /**
+   * A subclass of CanvasLayer which provides support for user interaction with its contained objects.
+   */
+  class InteractionLayer extends CanvasLayer {}
+}
