+++ src/foundry/client/pixi/layers/controls/index.d.ts
@@ -1,4 +1,5 @@
 import "./cursor";
 import "./door";
 import "./hud";
+import "./layer";
 import "./ruler";
