+++ src/foundry/client/pixi/layers/effects/index.d.ts
@@ -1,1 +1,6 @@
-import "./particles";
+import "./weather/particles";
+import "./background-effects";
+import "./coloration-effects";
+import "./illumination-effects";
+import "./visibility";
+import "./weather-effects";
