+++ src/foundry/client/pixi/layers/effects/weather-effects.d.ts
