+++ src/foundry/client/pixi/layers/effects/weather/particles/effect.d.ts
