+++ src/foundry/client/pixi/layers/effects/weather/particles/index.d.ts
