+++ src/foundry/client/pixi/layers/effects/weather/particles/leaves.d.ts
