+++ src/foundry/client/pixi/layers/effects/weather/particles/rain.d.ts
