+++ src/foundry/client/pixi/layers/effects/weather/particles/snow.d.ts
