+++ src/foundry/client/pixi/layers/grid/index.d.ts
@@ -1,4 +1,5 @@
 import "./grid";
 import "./hex";
 import "./highlight";
+import "./layer";
 import "./square";
