+++ src/foundry/client/pixi/layers/index.d.ts
@@ -1,15 +1,8 @@
+import "./base";
 import "./controls";
-import "./controls/index";
-import "./drawings";
 import "./effects";
-import "./effects/index";
 import "./grid";
-import "./grid/index";
-import "./lighting";
+import "./masks";
+import "./placeables";
 import "./map";
-import "./notes";
 import "./sight";
-import "./sounds";
-import "./templates";
-import "./tokens";
-import "./walls";
