+++ src/foundry/client/pixi/layers/masks/index.d.ts
@@ -1,0 +1,3 @@
+import "./depth";
+import "./occlusion";
+import "./vision";
