+++ src/foundry/client/pixi/layers/placeables/drawings.d.ts
