+++ src/foundry/client/pixi/layers/placeables/index.d.ts
@@ -1,0 +1,8 @@
+import "./drawings";
+import "./lighting";
+import "./notes";
+import "./sounds";
+import "./templates";
+import "./tiles";
+import "./tokens";
+import "./walls";
