+++ src/foundry/client/pixi/layers/placeables/lighting.d.ts
