+++ src/foundry/client/pixi/layers/placeables/sounds.d.ts
