+++ src/foundry/client/pixi/layers/placeables/templates.d.ts
