+++ src/foundry/client/pixi/layers/placeables/tiles.d.ts
@@ -1,0 +1,8 @@
+export {};
+
+declare global {
+  /**
+   * A PlaceablesLayer designed for rendering the visual Scene for a specific vertical cross-section.
+   */
+  class TilesLayer extends PlaceablesLayer<any> {}
+}
